backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.3.7
supabase>=2.0.2
langchain>=0.0.350
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.25.0
numpy>=1.26.0
tiktoken==0.5.2
